Renamed folders to lbdashboard
